feat: Display API connection status
Add a visual indicator to the footer that shows whether the application is connected to the API. This is achieved by checking for the presence of the API_KEY environment variable and displaying a connection status along with the last four characters of the key. This helps users quickly ascertain the operational status of the backend services.